refactor: Renamed Monsters to Cult (#272)
Changed all instances of "monsters" or "monster" to Cult, including
renamed cards. Also made some minor tweaks and grammar corrections to
strings here and there.

This may cause issues. We had to create "RenamedRoles.js" to cover for
roles that changed names. I changed the names of a few cards referencing
Monsters as well as the Monsters role folder to Cult, so I'm not sure
how it'll go.

#147