New role: Yandere (#489)
Taken from the Role Patrol list. Acts like a Serial Killer, falls in
love with a target during N1 and will die if the target dies. Target
does not know that there is a Yandere in love with them, they are not
alerted and will not die if the Yandere dies.