Feat: add Vice President, rework bride to make targets fall in love, rename bride to suitress (#669)
Vice President role from EM. Code ported from /v/illage and modified by
myself. There are some flaws with VP in terms of balancing and win
condition as outlined in the Role Patrol thread, but this is a 100%
faithful recreation of the role as it was on EpicMafia so expect further
changes later on.

Also did edits to Bride in same branch. Wedding rings now life-link the
proposer and actor upon acceptance, and Bride has been renamed Suitress
per Role Patrol spreadsheet.

---------

Co-authored-by: joyce <24848927+pikulet@users.noreply.github.com>