feat: cult/mafia jointing "benandante" role (#1004)
requested by at least 3 users

in normal circumstances, a mafia majority and a cult minority means the
mafia wins and the cult loses. the same is true vice-versa

many people have asked me for a special circumstances where the two
teams can joint. normally I don't support a joint between the
alignments, but for this I'm willing to make an exception

the goal is that if a Benandante is alive at the end of the game and
either the mafia or the cult wins, the other alignment will also win
alongside them

labeling this as a Draft because I'm not sure how to handle the
winCheck; modify the WinWithMafia and WinWithCult cards, or make a new
card?

---------

Co-authored-by: joyce <24848927+pikulet@users.noreply.github.com>