feat: wei boyang role (#1087)
gunrunner equivalent for the cult

intended effect: if a non-cult player is shot, they convert to cultist.
if a cult player is shot, they die

the way I did it is a little wonky. I had the gunshot actually heal the
target so that it would override the kill. there is probably a cleaner
way to do this right? don't want to run into unintended effects with a
gunshot that heals

---------

Co-authored-by: joyce <24848927+pikulet@users.noreply.github.com>