Fix for adding avatar to avi-less profile
I forgot to separate the check for determining whether the profile already had a profile, and whether the profile is getting updated with an avatar.
Before, it would only add an avatar to the profile that exists IF it already had a profile.
With the new change, it will upload the avatar whether or not it had an avatar before.