fix: polar bears meet together, rewritten (#1222)
rewritten version of an older attempt to make polar bears share a
meeting

not sure why i never thought of this. took out "multi" flag from polar
bear meeting and just gave them two separate visits.

the only flaw is that technically they can polarize the same person in
both meetings, but a smart player will recognize that that's a dumb
idea. there is a point where a technical flaw is actually just a skill
issue