update setup files and documentation (#1475)
I've been analyzing issues with environment setup including a bug
involved with login and loading the startup page. These issues are not
present in the production release but their presence in the local
environment has lead to issues contributing. The bug is still present
but I have documented the workaround in the `README.md`. On top of this
I've documented issues with running `sudo`.

I've also updated some of our scripts. In the root `package.json`, I've
updated `npm start` and `npm stop` to properly start and stop the docker
containers respectively. The `totalsetup.sh` and `cleanup.sh` scripts
were updated with comments and shebangs.

---------

Co-authored-by: Golbolco <Golbolco@gmail.com>