debug (games): height must be explicitly set for #root (#1586)
## How the issue occurred

While fixing the frontend code, it was observed that `height` was set
explicitly across several elements where it was not necessary.

Given that the codebase needs to be optimised, I went ahead and removed
all instances of explicitly set `height` properties that did not appear
to be necessary. This was problematic.

There is likely inner logic in the games components where the main chat
area inherits the height through a chain that leads to the `#root`
element.