Added Snake Charmer, Barber, Evil Twin, Dreamer from BOTC. Minor improvement to Banished roles. (#1667)
Barber, If a Barber dies A random Cult or Mafia player gets to switch
two players roles.

Snake Charmer added as Maiden, Each night a Maiden selects a player if
that player is mafia or cult, They switch roles with the Maiden.

Dreamer added as Investigator, Each choose a player to learn 2 roles.
One is the player's real role and one is a role of the opposite
alignment.

Evil Twin, At the start of the game an Evil Twin is assigned a Good
Twin. The Two Twins learn each others roles. If the Good Twin is
condemned Cult Wins.

Added a Message that appears if the base Banished Count is greater then
0. It tells players the base Banished count.

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>