No Dashii and Sweetheart from BOTC and Diviner bug fix. (#1668)
Added No Dashi as Miasma, Miasma blocks their neighbors (Skipping non
Village and Banished roles). If the blocked player is performing a
investigative action. They will flip the alignment of who ever they
visit. If they don't visit they flip a neighbors alignment.

Added Sweetheart, If a Sweetheart dies a random Village aligned player
will be blocked for the rest of the game. This blocking has the same
interaction as the Succubus/Miasma blocking.

Fixed Diviner crashing the site when checking a role with a modifier.