Added Godfather from BOTC. Banshed Improvements, Hider/Seeker/Invader improvements (#1696)
Hider and Seeker can kill the Invader. (This allows Invader to be used
more)
Hider will force a Seeker or Invader to spawn in closed setups.
Seeker will force a Hider or Invader to spawn in closed setups.
Invader will force a Hider or Seeker to spawn in closed setups.
The roles will replace a member of the Alignment they were from
(Village->Seeker,Mafia->Hider, etc)

Banished roles now work properly with unique roles.

Added Godfather from BOTC as Tormentor, Adds or Removes a Banished role.
Learns what Banished roles are in the current game. If a Banished role
dies in the day, May kill a player at night.

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>