Add Exorcist and Ghosthunter + Underworld (Display does not work) (#1716)
Added Exorcist, Each night choose a dead player, That player is
exorcised, They can't use any graveyard abilities, Be revived, or be
targeted.

added Ghostbuster,  Is the same as Exorcist but is mafia.

Players who veg are exorcised.

The Underworld only shows in side menu if a player vegs and will not
shows during the reviews. It will also sometimes show to the Exorcised
Player. (I do not know how to fix it)

The roles and vegging did function correctly in testing.

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>