Fixes for guiser interaction with armor, alerts (#1792)
For: https://github.com/UltiMafia/Ultimafia/issues/22

- Fixed guiser going thru armor immunity. It was the same bug in nature
as the turncoat bug from
https://github.com/UltiMafia/Ultimafia/pull/1782 so I tried a different
approach to fixing it this time. The dominates check for identity
stealer should no longer cause immune events.
- Fixed alerts being given to the wrong player when guising occurred;
swapIdentity now additionally swaps the recipients of alerts that are
presently queued.
- Added some guiser tests. The sockets.js change supports this because
it's a necessary function for swapIdentity.

Tested the alerts issue locally by having the gunsmith give a gun to a
cop then verifying that when they were guised, the dead player received
the alert for the gun and not the guiser.

Co-authored-by: Golbolco <Golbolco@gmail.com>