Fixed bug where guiser vegged if target kill vegged (#1796)
This picture shows the bug where the cop got kicked for not voting but
the guiser got vegged instead because they stole the identity of the
cop:

![firefox_96IGw036Y9](https://github.com/user-attachments/assets/94a7f770-130f-4f24-94d1-16d1f302423c)

The cop is dead even though the veg didn't target them because they got
killed normally by mafia.

This happens because identity steal occurs before the queued veg action.
When I tried to reproduce the bug after I made my changes, the expected
behavior of only the cop vegging occurred.

I don't know if this is the best possible fix for this bug but it feels
like a safe resolution to me.