Added user miniprofile and pronouns (#1811)
Resolves #1517 and #1518

Miniprofile appears whenever you hover over the panel of a user inside a
game. You can click on the user's panel to "lock" the miniprofile in
place. Clicking on the user's name inside the miniprofile takes you the
user's profile. Pronouns are conditionally displayed if set. Pie chart
is displayed using the same mechanism from the profile page. The
contents of the miniprofile are populated from a REST request that
happens once per user in the game per refresh of your page.

Pronouns are set from your profile, they default to the empty string and
your profile has a default message in place of your pronouns if they are
still empty. You can edit them exactly like your bio - the limit is 15
characters (it can be increased but it'd need to be tweaked to play nice
with the miniprofile). Profanity is filtered in the same way that the
bio is.

I also tested anonymous games to make sure I didn't break those - the
miniprofile doesn't appear as normal because noLink is set to true.

Screenshots:
### Appearance of miniprofile for brand new user:

![image](https://github.com/user-attachments/assets/0a606bef-0d97-40cd-b6e3-d53283b1703f)

### Appearance of miniprofile in hover state:

![image](https://github.com/user-attachments/assets/f10e6e04-62c4-4a7c-b00c-849b5513961d)

### Appearance of miniprofile in "locked" state:

![image](https://github.com/user-attachments/assets/78284a85-04fd-493d-8779-0099effece18)

### Appearance of profile for brand new user:

![image](https://github.com/user-attachments/assets/6678755f-93e6-4cf9-aaa6-ec1f33144912)

### Appearance of profile for established user:

![image](https://github.com/user-attachments/assets/9af7213d-f22b-4759-b8a0-7543068ec2c2)