Added kudos and karma, resolves #1509 (#1826)
### Appearance

Normal profile:

![image](https://github.com/user-attachments/assets/e9576498-4341-407f-b535-08c9fcde67e6)

Cramped profile:

![image](https://github.com/user-attachments/assets/4a55caa9-8e95-496e-ad7c-640884a56675)

Miniprofile:

![image](https://github.com/user-attachments/assets/05464d86-ec3d-4fc4-bb39-ca47755339bd)

### Karma

Happens very similarly to upvoting or downvoting a comment, except it's
tied to a user. Every single vote a user does for another is recorded
separately in the database, but votes are not visible to other users.

### Kudos

When a game is ranked or competitive, a kudos can be voted on to be
awarded in the postgame meeting. When the lobby closes (after all
players leave or 2 minutes, whichever comes first), the vote is
calculated and awards it to the winner.

Ties will be randomly awarded between the highest scorers, and games
with no kudos votes will give no kudos. You can see who received the
kudos by reviewing the game after it closes.

### Misc

- Fixed NameWithAvatar always querying for the user's profile, this is
fixed to only happen when it's needed in games now for the miniprofile
- Refactored miniprofile a bit
- Right justified the in love with section as the left side of the
profile was starting to get a little cramped