Added archived games functionality (#1963)
For #1809

Added a button to the game review screen where users, if they've
purchased the capability from the shop, can archive a game. Archived
games are exempt from expiration and appear on the user's profile. Users
can "unarchive" games from their profile. Unarchiving a game that is
beyond its expiration date will cause it to be deleted on the next
sweep.

Eventually I would like to add the ability to edit the description of an
archived game, but I'm not doing it for the first swing at it.


![image](https://github.com/user-attachments/assets/1ad666fb-78a4-43ed-90d2-1614ee64d3dc)