Added leaderboard endpoint (#1983)
To test:

Update mongoDB container with new mount (applies if using EZ setup):
`docker compose up -d`

Add leaderboard users (generate-leaderboard-users.js has a line at the
bottom to let you reverse it):
```
docker exec -it mongodb mongosh
load("/scripts/dev-login.js")
load("/scripts/generate-leaderboard-users.js")
```

You should now receive a list of users per field from the endpoint when
you hit the leaderboards page:

![image](https://github.com/user-attachments/assets/b641f587-c6a3-4a76-bc6e-9613c2e57e26)

I didn't integrate the response data with the page, someone will still
need to do that.

To add other fields, we will want to create a new field that's updated
when the value being tracked is updated, e.g. when an achievement is
earned a user's achievmentCount is set to the length of the achievements
list. A user's mafia winrate would be calculated at the end of a game,
at the same time that stats are updated.

That being said, I think that elo would be a preferred leaderboard stat
to winrate? For whenever we implement that