Reaper and Mole Roles + Humble, Suspect, Shady, bug fixes  (#1991)
Humble Now works with Shady, Suspect, Respected, etc

Braggart and Tele work better with Mi-Go and Philosopher 

Reaper- If you say "I claim Reaper and Choose (Player Name)" within the
first minute of the day.
and that player survives until the end of the day, That player's team
wins.
If you use your own name nothing happens.

Mole- Village - Becomes a Mafia or Cult role and counts towards their
majority. (Attends the Mafia/Cult if they have a non-lone role)
Once per game the Mafia or Cult can guess who the mole is even if dead.
On a correct guess mafia or cult win.
(The mole knows that they are the mole).

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>