new command /changeSetup (#2047)
Do /changeSetup {Setup ID} to change the current setup. 
Only host can do it. (Host player not host role)
Cannot change to a setup lower then current player count.
Cannot change to a setup of a different game type.
Cannot change in ranked/comp games for now

Bugs: 
Players have to refresh to see the new setup.
Setup will show as original setup in lobby browser until the game is
complete.
Probably other bugs aswell.

This can be used as a base to make my previous suggestion of setup
collections or we can be lazy and just leave it like this.