More Assassin Updates (everyone loves Assassin Updates) + Jazzman Updates (No one loves Jazzman) (#2270)
Role Sharing can now only be done within a room.
Talking and Voting Dead work in rooms.

Various roles now work in Assassin roles.

Added Mastermind from 2 Rooms and a Boom as a Special Interaction for
Fool, If an Assassin is present, a Fool will win at the end of the game
if they were elected Leader of 2 different Rooms during the game.

Added Minion from 2 Rooms and a Boom as Special Interaction for
Executioner, If an Assassin is present, an Executioner will win at the
end of the game if they are never in a room when the room leader
changes.

Added Bomb-Bot from 2 Rooms and a Boom as Special Interaction for
Vengeful Spirit, If an Assassin is present, a Vengeful Spirit will win
if the Assassin kills them and the Assassin does not kill any
Presidents, Senators, or Vital Village roles.

Made Special Interactions show on role pages.

Jazzman- now makes everyone delirious when any player is condemned.
Jazzman can now use Loyal, Disloyal, Simple, etc. (Makes the delirium
only happen when a player the modifier allows is condemned)
Old Jazzman = Jazzman (Disloyal);

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>