Bug Fixes + Dentist into Mafia role + Poet Changes (#2351)
Epilogue No Longer in every game.
Fixed Epilogue breaking Dusk roles.
Independents can win if a Ghost guess a word when a poet is in the game.
Removed Village players being able to Guess poet during Dusk.
Allowed living players to guess Poet during Epilogue. 
Fixed Crashes with Mail-In-ballots, Paralyzer, and Whistleblower, 
Allowed Mail-In-ballots, Paralyzer, and Whistleblower to work with
Assassin
 
Poet no longer guesses the word when condemned and appears as Villager
on death.

Dentist is now a mafia role.

---------

Co-authored-by: SawJester <SawJester@users.noreply.github.com>